AddLogin

+++ Account.Data/Entities/Account.cs
@@ -6,7 +6,7 @@ namespace Account.Data.Entities
     {
         public Guid AccountId { get; set; }
         public DateTime OpenDate { get; set; }
-        public double Balance { get; set; }
+        public int Balance { get; set; }
         public Guid CustomerId { get; set; }
         public virtual Customer Customer { get; set; }
     }
